LayoutTests/imported/w3c/web-platform-tests/service-workers/service-worker/fetch-request-css-images.https-expected.txt



PASS Verify FetchEvent for css image (backgroundImage). 
FAIL Verify FetchEvent for css image (shapeOutside). assert_equals: credentials of CSSImage load (url:https://localhost:9443/service-workers/service-worker/resources/dummy?test1516682573070 type:shapeOutside) must be same-origin. expected "same-origin" but got "include"
PASS Verify FetchEvent for css image-set (backgroundImage). 
FAIL Verify FetchEvent for css image-set (shapeOutside). assert_equals: credentials of CSSImageSet load (url:https://localhost:9443/service-workers/service-worker/resources/dummy?test1516682573028 type:shapeOutside) must be same-origin. expected "same-origin" but got "include"
